dev-13777 counterpart selector in its own file
